Reduce Disk Writes With noatime, nodiratime noatime, nodiratime are disk mount options to disable access time information updates onto every file or directory the system access. By using these options, disk writing activity can be reduce significantly.
    noatime - Disable access time update of files in HDD.
    nodiratime - Disable access time update of directories in HDD
****************************************************************************************

[Explanation]
Turns out our save files, for each of our files, a parameter called "atime" or "access time". Total, that each time you access (read) a file, the value "atime" is updated. So we know which was the last time we agreed / we read a particular file.

Of course, this is an unwanted behavior: It turns out that we do read each disk, we need to write something to disk! (Update the value "access time"). This seems absurd! To read we have to write.

Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>