i2c-qup: don't ignore the last byte of a message
Ignoring the last byte of the I2C write message when
the I2C transfer has the multiple messages with
write and read from different slave addresses. This
leads to i2c master pulls the clock line low and
also client misses one byte of data from the
write message.

This patch allows the last byte of the write message
in this particular case also.

Change-Id: I433479e95c5f167dce56d57e5c9ca1c65312ff68
Signed-off-by: Kiran Gunda <kgunda@codeaurora.org>
Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>